scripts/requirements.txt
flask==3.0.2
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
gunicorn==21.2.0 